requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
drf-yasg
django-extensions
jupyter
ipython
drf-spectacular